+++ lib/Array.ts
@@ -8,7 +8,7 @@ export namespace ArrayUtils {
    * A typesafe way for removing all null/undefined values from an array.
    */
   export const removeOptionals = <T>(arr: (T | null | undefined)[]) => {
-    return arr.reduce((acc, curr) => (curr ? [...acc, curr] : acc), [] as T[])
+    return arr.reduce((acc, curr) => (curr ? [...acc, curr] : acc), [])
   }
 
   /**
