+++ lib/location/search/index.ts
@@ -1,2 +1,3 @@
 export * from "./History"
-export * from "./Client"
+export * from "./TextSearch"
+export * from "./Result"
